Support for xattr properties.
Each property is named getxattr in the namespace defined by the
NGX_HTTP_DAV_EXT_XML_NS_XATTR macro.  The actual attribute to retrieve
is put in the name attribute in the same namespace.

The getxattr elements in the response have a type attribute which can
be either "text/plain" in which case the value is literally contained
in the element or "application/base64" if the value is not printable
and has been encoded in Base64.

The dav_ext_getxattr configuration directive can be used to set a
regular expression that xattr names have to match in order to be
exposed as properties.