Fix fetching xattr attributes when propfinding a collection
If one child or the collection itself does not have the attribute, then it
would never be requested again on the following children.

The most common case is propfinding a collection which does not have any
xattr, so the xattr of the children would not be listed.

Fix that by making props_not_found not own the props that are not found,
so that the original list of properties to retrieve is unchanged.